% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateotl.R
\name{interpolateOTL}
\alias{interpolateOTL}
\title{interpolateOTL}
\usage{
interpolateOTL(profiles, criteria, precision)
}
\arguments{
\item{profiles}{Dataframe - vertical profiling data. Fields must include c("Lake", "Date", "Station", "Depth_ft", "Parameter", "Value", "Units")}

\item{criteria}{List - DO and Temperature thresholds. Accepts list output from setCriteria()}

\item{precision}{numeric precision for vertical interpolation. 0.1 foot recommended.}
}
\description{
interpolateOTL
}
