% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmultiple.R
\name{plotMultiple}
\alias{plotMultiple}
\title{Title}
\usage{
plotMultiple(Bands, stations, years, criteria)
}
\arguments{
\item{Bands}{Dataframe - ouput of interpolateOTL}

\item{criteria}{List - output of setCriteria}

\item{station}{Character Vector - names of stations within Bands to plot}

\item{year}{Numeric Vector - years to plot}
}
\description{
Title
}
