% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsingle.R
\name{plotSingle}
\alias{plotSingle}
\title{Title}
\usage{
plotSingle(Bands, station, year, criteria)
}
\arguments{
\item{Bands}{Dataframe - ouput of interpolateOTL}

\item{station}{Character - name of station within Bands}

\item{year}{Numeric - year to plot}

\item{criteria}{List - output of setCriteria}
}
\description{
Title
}
