% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setcriteria.R
\name{setCriteria}
\alias{setCriteria}
\title{setCriteria}
\usage{
setCriteria(minDO, maxTemperature)
}
\arguments{
\item{minDO}{minimum Dissolved Oxygen criteria}

\item{maxTemperature}{maximum Dissolved Oxygen criteria}
}
\description{
setCriteria
}
